Split FlockBase into Drone and FlockBase
